requirements.txt
# Web Scraper/Crawler
Scrapy

#Twitter API
twitter
tweepy

#Configurations
python-dotenv